+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * [#550](https://github.com/mozilla-rally/rally-core-addon/pull/550): Do not use the extension ID as the schema namespace. Re-use the "keyId" argument for this since it is redundant.
 * [#532](https://github.com/mozilla-rally/rally-core-addon/pull/532): Log the payload to console when data collection is disabled, for QA purposes.
-
+* [#547](https://github.com/mozilla-rally/rally-core-addon/pull/547): Update copy by changing "partner" to "collaborator".
 # v1.2.0 (2021-04-12)
 
 [Full changelog](https://github.com/mozilla-rally/core-addon/compare/v1.1.0...v1.2.0)